+++ source/CameraSystem.py
@@ -1,9 +1,9 @@
-from numberDetection import DetectionNumber as dn
+from DetectionNumber import DetectionNumber
 
 
 class CameraSystem:
     def __init__(self):
-        self.detection_number = dn.DetectionNumber()
+        self.detection_number = DetectionNumber()
 
     def main(self):
         pass
