+++ source/DetectionNumber/clipping_number.py
