+++ source/DetectionNumber/training_scripts/original/6.jpg
